+++ js/ui.modal.js
@@ -20,15 +20,22 @@ define(function(require, exports, module) {
     var Modal = function(element, options) {
         this.options = $.extend({}, Modal.DEFAULTS, options || {});
         this.$element = $(element);
+
+        if (!this.$element.attr('id')) {
+            this.$element.attr('id', UI.utils.generateGUID('am-modal'))
+        }
+
         this.isPopup = this.$element.hasClass('am-popup');
         this.active = this.transitioning = null;
 
         this.events();
     };
 
-
-    // Set template tabindex to tigger onfocus on div
     Modal.DEFAULTS = {
+        className: {
+            active: 'am-modal-active',
+            out: 'am-modal-out'
+        },
         selector: {
             modal: '.am-modal',
             active: '.am-modal-active'
@@ -37,7 +44,8 @@ define(function(require, exports, module) {
         onConfirm: function() {
         },
         onCancel: function() {
-        }
+        },
+        duration: 300 // must equal the CSS transition duration
     };
 
     Modal.prototype.toggle = function(relatedElement) {
@@ -46,6 +54,7 @@ define(function(require, exports, module) {
 
     Modal.prototype.open = function(relatedElement) {
         var $element = this.$element,
+            options = this.options,
             isPopup = this.isPopup;
 
         if (this.transitioning || this.active) return;
@@ -58,17 +67,13 @@ define(function(require, exports, module) {
 
         $element.trigger($.Event('open:modal:amui', {relatedElement: relatedElement}));
 
-        // trigger reflow;
-        var clientLeft = $element[0].clientLeft;
+        dimmer.open($element);
 
-        dimmer.open();
+        $element.show().redraw();
 
-        $element.show();
+        !isPopup && $element.css({marginTop: - parseInt($element.height() / 2, 10) + 'px'});
 
-        !isPopup && $element.redraw()
-            .css({marginTop: -$element.height() / 2 + 'px'});
-
-        $element.removeClass('am-modal-out').addClass('am-modal-active');
+        $element.removeClass(options.className.out).addClass(options.className.active);
 
         this.transitioning = 1;
 
@@ -79,13 +84,14 @@ define(function(require, exports, module) {
 
         if (!supportTransition) return complete.call(this);
 
-        $element.one(supportTransition.end, $.proxy(complete, this));
+        $element.one(supportTransition.end, $.proxy(complete, this)).emulateTransitionEnd(options.duration);
     };
 
     Modal.prototype.close = function(relatedElement) {
         if (this.transitioning || !this.active) return;
 
         var $element = this.$element,
+            options = this.options,
             isPopup = this.isPopup,
             that = this;
 
@@ -95,23 +101,20 @@ define(function(require, exports, module) {
 
         var complete = function() {
             $element.trigger('closed:amui:modal');
-            isPopup && $element.removeClass('am-modal-out');
+            isPopup && $element.removeClass(options.className.out);
             $element.hide();
             this.transitioning = 0;
         };
 
-        $element.removeClass('am-modal-active').addClass('am-modal-out');
+        $element.removeClass(options.className.active).addClass(options.className.out);
 
         if (!supportTransition) return complete.call(this);
 
         $element
             .one(supportTransition.end, $.proxy(complete, this))
-            .emulateTransitionEnd(300);
+            .emulateTransitionEnd(options.duration);
 
-        // hide dimmer when all modal is closed
-        if (!$body.find(Modal.DEFAULTS.selector.active).length) {
-            dimmer.close();
-        }
+        dimmer.close($element);
 
         this.active = false;
     };
@@ -188,6 +191,3 @@ define(function(require, exports, module) {
 
     module.exports = Modal;
 });
-
-// TODO: 1. parse tpl?
-//       2. firefox transitionend 关闭时没有动效 参考： http://getuikit.com/docs/modal.html